requirements.txt
configparser==5.0.1
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.1
Pillow==8.1.2
pyparsing==2.4.7
PySide6==6.0.0
python-dateutil==2.8.2
shiboken6==6.0.0
six==1.16.0
utils==1.0.1
